Isolate components and remove prefixes from files
libecho-ui is renamed to libecho-client to become a more general module,
and our extension interface is isolated in its own module to provide our
core library with some breathing room. You may also notice that we
remove the namespace prefix from files. This should, in turn, make
inclusion more natural:

<libecho/client/window.h> vs <libecho/client/echo-client-window.h>.

On an additional note, we now bundle resource files (.c) to compile
distribution should glib-compile-resource be missing. This is mostly
targeted at CI, and not for end-users, but both sides can take advantage
of it.